+++ src/problem/Problems.kt
@@ -6,6 +6,7 @@ const val BinCover = "bin_cover"
 const val BinPacking = "bin_packing"
 const val Clique = "clique"
 const val CliqueCover = "clique_cover"
+const val DominatingSet = "dominating_set"
 const val Interval = "interval"
 const val KCut = "k_cut"
 const val Knapsack = "knapsack"
@@ -26,6 +27,7 @@ val Creator: Map<String, (String, Int) -> Problem?>  = mapOf(
 	BinPacking to ::newBinPacking,
 	Clique to ::newClique,
 	CliqueCover to ::newCliqueCover,
+	DominatingSet to ::newDominatingSet,
 	Interval to ::newInterval,
 	KCut to ::newKCut,
 	Knapsack to ::newKnapsack,